Move CommentTree.svelte and CommentTree.stories.ts into CommentTree directory
